Rename app.py into data_pipeline.py
